backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Binance API
python-binance==1.0.19

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Utils
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# Scheduling
apscheduler==3.10.4

# CORS
starlette==0.35.1
